refactor: generalise per-chromosome summary stats ingestion
